added practical2/practical6

+++ practical_1/program_1.py
@@ -1,8 +1,8 @@
-score_1 = int(input("Enter Test Score 1: "))
-score_2 = int(input("Enter Test Score 2: "))
-score_3 = int(input("Enter Test Score 3: "))
-score_4 = int(input("Enter Test Score 4: "))
-score_5 = int(input("Enter Test Score 5: "))
+score_1 = float(input("Enter Test Score 1: "))
+score_2 = float(input("Enter Test Score 2: "))
+score_3 = float(input("Enter Test Score 3: "))
+score_4 = float(input("Enter Test Score 4: "))
+score_5 = float(input("Enter Test Score 5: "))
 
 avg = (score_1 + score_2 + score_3 + score_4 + score_5)/5
